+++ src/applications/vaos/tests/services/mocks/index.unit.spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { generateMockSlots } from '../../../services/mocks';
+import generateMockSlots from '../../../services/mocks/var/slots';
 
 describe('calendar mock data', () => {
   it('should generate 300 mock slots', () => {
